Use LMQ's new outgoing-only option for reachability testing
Use connect_remote was causing significant problems with the default
pubkey-based routing IDs because the new outgoing connections for
reachability would end up replacing the existing SN-to-SN connection due
to reusing the same id.

This commit includes an update to lokimq to specify that a send/request
should be done only on outgoing connections, without needing to
open and close new outgoing connections when there is an existing
outgoing SN connection to the remote being tested.