Add heuristic to significantly avoid bootstrap use
The bootstrapping code was added because, long ago, oxend didn't have
ed/x keys and ip addresses after a restart, so on a SN restart it could
take an hour before we had a full view of the network.

This isn't the case anymore: ed/x/ip/etc. fields are persistent, so the
bootstrap code here is really only necessary for a fresh service node.

This commit thus adds a heuristic to only go to the bootstrap node if
our first snode list from oxend is missing data for more than 3% of the
service nodes; otherwise we just trust our local oxend result so that we
should, on an up-to-date service node, almost never need to contact the
bootstraps.

This also tweaks when we update local swarm data when bootstrapping so
that we recognize ourself as being active right away; the change to
pushed block notifications broke this by adding a delay to the next
arriving block (I think the old version didn't have this issue because
it hammered oxend once/second and so there would have only been a 1s
delay after bootstrap until the next rpc request).